requirements.txt
asgiref==3.7.2
Django==5.0.2
django-extensions==3.2.3
djangorestframework==3.14.0
numpy==1.26.4
pandas==2.2.0
python-dateutil==2.8.2
pytz==2024.1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
wheel==0.42.0
